Refactor as a Vite app with npm dependencies and modular source
Converts the static-file project (CDN scripts, a single monolithic
js/script.js) into a Vite build with real npm dependencies pinned to
the exact versions previously loaded from CDN: maplibre-gl@5.24.0,
jquery@2.1.4, bootstrap@3.3.6, plus the individual turf packages
actually used (@turf/circle, @turf/union, @turf/simplify,
@turf/helpers) rather than the full @turf/turf bundle.

Source is split into src/ modules along the app's actual seams:
constants, the ScratchLayer WebGL class, the border vector layer (and
its Web Worker), OpenFreeMap labels, an orchestration module that
wires raster/scratch/border/labels together and handles swap, drag
interaction handling, and jQuery/Bootstrap UI wiring.

data/label-layers.json becomes a bundled import instead of a runtime
fetch, which also removes the fetch-vs-style-load race condition that
existed before (only the style-readiness check remains now).

Fixed a real bug surfaced while wiring up Bootstrap: static `import`
statements are always hoisted above other code in a module, so
`window.jQuery = $` (needed because Bootstrap 3's plugin expects a
global) was executing after Bootstrap's own import had already run
and thrown "Bootstrap's JavaScript requires jQuery". Fixed with a
dynamic import() plus top-level await, which runs in real statement
order.

Also removes genuinely dead files left over from before the MapLibre
migration (orphaned Leaflet geosearch plugin files, a stray .DS_Store)
that were referenced nowhere.

Verified via dev server, production build, and preview server -- all
core functionality (scratch, border, swap, labels, mode switching, the
About modal) works identically to before.
