+++ .gitignore
@@ -1,1 +1,3 @@
 /scripts
+node_modules
+dist
