+++ index.html
@@ -4,9 +4,6 @@
 <head>
   <meta charset="UTF-8">
   <title>Urban Scratchoff</title>
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/maplibre-gl@5.24.0/dist/maplibre-gl.css" />
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css">
-  <link rel="stylesheet" href="css/styles.css">
 </head>
 
 <body>
@@ -197,12 +194,7 @@
     ga('send', 'pageview');
 
   </script>
-  <script src="https://cdn.jsdelivr.net/npm/maplibre-gl@5.24.0/dist/maplibre-gl.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/@turf/turf@7.4.0/turf.min.js"></script>
-  <script src="https://code.jquery.com/jquery-2.1.4.js"></script>
-  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script>
-  <script type="js/routers/router.js"></script>
-  <script src="js/script.js"></script>
+  <script type="module" src="/src/main.js"></script>
 
 </body>
